Added sparse matrices: - Supported format: coordinate and compressed row storage - Matrix types: GeCoordMatrix, GeCRSMatrix, SyCoordMatrix, SyCRSMatrix - Examples: gecoord-example.cc, sycoord-example.cc
In FEM you usually need the coordinate storage during assembling the stiffness matrix.  For solving the resulting system of linear equations you need to convert the matrix to a compressed format first.  The examples show how to do this tasks.

We also support the conversion from sparse matrices to GeMatrix (or SyMatrix) mostly for debugging purposes.